requirements.txt
Langchain
ipykernel
python-dotenv
Langchain_community
pypdf
bs4
arxiv
pymupdf
langchain_text_splitters
Langchain_openai
chromadb
sentence-transformers
Langchain-huggingface
faiss-cpu
langchain_chroma
streamlit
validators
youtube_transcript_api
unstructured
pytube
google-generativeai
youtube-transcript-api